move get project secret provider to crud/secrets.py
